Move inner sub-components into MapSelector
