+++ frontend/src/features/properties/selector/components/SelectedPropertyHeaderRow.tsx
