+++ frontend/src/features/properties/selector/components/SelectedPropertyRow.test.tsx
