+++ frontend/src/features/properties/selector/components/SelectedPropertyRow.tsx
